+++ src/main/java/net/fabricmc/mappingio/MappingWriter.java
@@ -27,6 +27,7 @@ import org.jetbrains.annotations.Nullable;
 import net.fabricmc.mappingio.format.MappingFormat;
 import net.fabricmc.mappingio.format.enigma.EnigmaDirWriter;
 import net.fabricmc.mappingio.format.enigma.EnigmaFileWriter;
+import net.fabricmc.mappingio.format.jobf.JobfFileWriter;
 import net.fabricmc.mappingio.format.proguard.ProGuardFileWriter;
 import net.fabricmc.mappingio.format.simple.RecafSimpleFileWriter;
 import net.fabricmc.mappingio.format.srg.JamFileWriter;
@@ -65,6 +66,7 @@ public interface MappingWriter extends Closeable, MappingVisitor {
 		case TSRG_2_FILE: return new TsrgFileWriter(writer, true);
 		case PROGUARD_FILE: return new ProGuardFileWriter(writer);
 		case RECAF_SIMPLE_FILE: return new RecafSimpleFileWriter(writer);
+		case JOBF_FILE: return new JobfFileWriter(writer);
 		default: return null;
 		}
 	}