+++ src/main/java/net/fabricmc/mappingio/format/MappingFormat.java
@@ -122,6 +122,15 @@ import org.jetbrains.annotations.Nullable;
  *     <td>-</td>
  *     <td>-</td>
  *   </tr>
+ *   <tr>
+ *     <td>JOBF</td>
+ *     <td>-</td>
+ *     <td>src</td>
+ *     <td>-</td>
+ *     <td>-</td>
+ *     <td>-</td>
+ *     <td>-</td>
+ *   </tr>
  * </table>
  */
 // Format order is determined by importance to Fabric tooling, format family and release order therein.
@@ -186,7 +195,12 @@ public enum MappingFormat {
 	/**
 	 * Recaf's {@code Simple} mapping format, as specified <a href="https://github.com/Col-E/Recaf/blob/e9765d4e02991a9dd48e67c9572a063c14552e7c/src/main/java/me/coley/recaf/mapping/SimpleMappings.java#L14-L23">here</a>.
 	 */
-	RECAF_SIMPLE_FILE("Recaf Simple file", "txt", false, true, false, false, false, true);
+	RECAF_SIMPLE_FILE("Recaf Simple file", "txt", false, true, false, false, false, true),
+
+	/**
+	 * The {@code JOBF} mapping format, as specified <a href="https://github.com/skylot/jadx/blob/2d5c0fda4a0c5d16207a5f48edb72e6efa7d5bbd/jadx-core/src/main/java/jadx/core/deobf/DeobfPresets.java">here</a>.
+	 */
+	JOBF_FILE("JOBF file", "jobf", false, true, false, false, false, true);
 
 	MappingFormat(String name, @Nullable String fileExt,
 			boolean hasNamespaces, boolean hasFieldDescriptors,