+++ src/test/java/net/fabricmc/mappingio/read/ValidContentReadTest.java
@@ -123,6 +123,13 @@ public class ValidContentReadTest {
 		checkHoles(format);
 	}
 
+	@Test
+	public void jobfFile() throws Exception {
+		MappingFormat format = MappingFormat.JOBF_FILE;
+		checkDefault(format);
+		checkHoles(format);
+	}
+
 	private VisitableMappingTree checkDefault(MappingFormat format) throws Exception {
 		VisitableMappingTree tree = new MemoryMappingTree();
 		MappingReader.read(TestHelper.MappingDirs.VALID.resolve(TestHelper.getFileName(format)), format, tree);