+++ src/test/java/net/fabricmc/mappingio/write/WriteTest.java
@@ -103,6 +103,11 @@ public class WriteTest {
 		check(MappingFormat.RECAF_SIMPLE_FILE);
 	}
 
+	@Test
+	public void jobfFile() throws Exception {
+		check(MappingFormat.JOBF_FILE);
+	}
+
 	private void check(MappingFormat format) throws Exception {
 		dogfood(validTree, dir, format);
 		dogfood(validWithHolesTree, dir, format);